#101 SG018 Rozliczenie bonow: Rozliczenie - rozliczenie TI - zmiana z BigDecimali na Integery
git-svn-id: http://svn.sodexo.com.pl/svn/gryf/trunk@400 a1b3532b-92c3-dd42-b132-913ad673e08d

+++ gryf-common/src/main/java/pl/sodexo/it/gryf/common/dto/publicbenefits/electronicreimbursements/CalculationChargesParamsDto.java
@@ -20,19 +20,18 @@ public class CalculationChargesParamsDto {
 
     @Getter
     @Setter
-    private BigDecimal productInstanceForHour;
+    private Integer productInstanceForHour;
 
     @Getter
     @Setter
-    private BigDecimal maxProductInstance;
+    private Integer maxProductInstance;
 
     @Getter
     @Setter
-    private BigDecimal usedProductsNumber;
+    private Integer usedProductsNumber;
 
     @Getter
     @Setter
     private BigDecimal productValue;
-
 
 }
